requirements.txt
SQLAlchemy~=2.0.23
alembic~=1.13.0
Pillow~=10.1.0
numpy~=1.26.2
opencv-contrib-python~=4.8.1.78
jsonpickle~=3.0.2
scikit-learn~=1.3.2
email-validator~=2.1.0.post1